+++ pumpkin-lib/src/propagators/cumulative/time_table/mod.rs
@@ -70,9 +70,11 @@
 //! 149–157.
 
 mod time_table_per_point;
+mod time_table_per_point_incremental;
 mod time_table_util;
 
 pub use time_table_per_point::*;
+pub use time_table_per_point_incremental::*;
 
 #[cfg(doc)]
 use crate::propagators::cumulative::time_table::time_table_util::*;