+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Added disjunctive scheduling example through a difference logic model.
 * Added cumulative propagator(s)
   * Added propagator utilising time-table reasoning per time point
+  * Added incremental version of a propagator utilising time-table reasoning per time point
 * Extended collection of statistics to include conflict size, learned clause length and backtrack amount
 * Added logging of statistics using `info!` when using the `OptimisationSolver` 
 * Refactored the CP and propositional trails, and extracted a trail abstraction.